Remove very old pickle compatibility code modifying old atg package names
The code was meant to support opening some old datasets that contained
pickled data with pre-petastorm names.
